videoGen/requirements.txt
av==13.1.0
beautifulsoup4==4.14.3
certifi==2026.1.4
click==8.3.1
cloudinary==1.44.1
cloup==3.0.8
decorator==5.2.1
glcontext==3.0.0
isosurfaces==0.1.2
manim==0.19.1
ManimPango==0.6.1
mapbox_earcut==2.0.0
markdown-it-py==4.0.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
networkx==3.4.2
numpy==2.2.6
pillow==12.1.0
pycairo==1.29.0
pydub==0.25.1
pyglet==2.1.13
pyglm==2.8.3
Pygments==2.19.2
python-dotenv==1.2.1
rich==14.3.2
scipy==1.15.3
screeninfo==0.8.1
six==1.17.0
skia-pathops==0.9.1
soupsieve==2.8.3
srt==3.5.3
svgelements==1.9.6
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
watchdog==6.0.0
